--- a/frontend_tests/node_tests/stream_data.js
+++ b/frontend_tests/node_tests/stream_data.js
@@ -304,6 +304,7 @@ run_test('is_active', () => {
 
     page_params.demote_inactive_streams =
         settings_display.demote_inactive_streams_values.automatic.code;
+    stream_data.set_filter_out_inactives();
 
     sub = {name: 'pets', subscribed: false, stream_id: 111};
     stream_data.add_sub('pets', sub);
@@ -334,6 +335,7 @@ run_test('is_active', () => {
 
     page_params.demote_inactive_streams =
         settings_display.demote_inactive_streams_values.always.code;
+    stream_data.set_filter_out_inactives();
 
     sub = {name: 'pets', subscribed: false, stream_id: 111};
     stream_data.add_sub('pets', sub);
@@ -361,6 +363,7 @@ run_test('is_active', () => {
 
     page_params.demote_inactive_streams =
         settings_display.demote_inactive_streams_values.never.code;
+    stream_data.set_filter_out_inactives();
 
     sub = {name: 'pets', subscribed: false, stream_id: 111};
     stream_data.add_sub('pets', sub);