--- a/static/js/bundles/app.js
+++ b/static/js/bundles/app.js
@@ -31,6 +31,7 @@ import "../search_util.js";
 import "../keydown_util.js";
 import "../lightbox_canvas.js";
 import "../rtl.js";
+import "../lazy_set.js";
 import "../dict.ts";
 import "../scroll_util.js";
 import "../components.js";