--- a/static/js/server_events_dispatch.js
+++ b/static/js/server_events_dispatch.js
@@ -409,6 +409,7 @@ exports.dispatch_normal_event = function dispatch_normal_event(event) {
         }
         if (event.setting_name === 'demote_inactive_streams') {
             stream_list.update_streams_sidebar();
+            stream_data.set_filter_out_inactives();
         }
         if (event.setting_name === 'dense_mode') {
             $("body").toggleClass("less_dense_mode");