@showell 2019-12-27
I tested this with about 500k subscribers, and then I put some timings around these lines:

~~~ js
702     const t = new Date().getTime();
703     populate_subscriptions(page_params.subscriptions, true, true);
704     populate_subscriptions(page_params.unsubscribed, false, true);
705     populate_subscriptions(page_params.never_subscribed, false, false);
706     const elapsed = new Date().getTime() - t;
707     console.info('subscriptions', elapsed);
~~~

* master: 670ms
* PR: 31ms
* PR without lazy-set optimization: 90ms

The big win here is getting away from `Dict`.  But the additional 60ms for making the set lazy seems worthwhile to me, as the code is pretty simple.