@showell 2019-12-27
I also added a couple more commits:

* remove N-squared calls to get number of subscribed streams
* make it fast to ignore unsubscribed streams