@andersk 2019-12-28
We’re using Babel now (#12838), so we may as well use modern JS like `class`es and arrow functions for new code, even though not much old code has been modernized yet.

I haven’t checked in detail, but does this run into a problem with `.subscribed` being mutated?