@showell 2019-12-28
Good catch on the `subscriber` mutate.  I added a test for that (as the first commit) and made sure it didn't break with my changes.