@timabbott 2019-12-28
This is awesome, thanks for doing this @showell!  I posted a few relatively minor comments; I'm fine with handling the `server_events_dispatch.js` change in a new issue if desired just to get the main work of this integrated sooner.