@showell 2019-12-28
I'm hoping we can merge this as is, since the 600ms reduction is a clear win.