@timabbott 2019-12-30
Merged, after tweaking the BinaryDict commit message and adding `streams:` prefixes on a few commits.  Thanks @showell!  I'm about to test-deploy this to chat.zulip.org.